1337 Cursus

+++ cpp/Module07/ex01/iter.hpp
@@ -4,7 +4,7 @@
 template <typename A, typename C>
 void	iter(A first, size_t b, C third)
 {
-	for (int i = 0; i < b; i++)
+	for (size_t i = 0; i < b; i++)
 	{
 		third((*first)[i]);
 	}
